fix version regex, move payload to fixure folder
